#32: Added more Entrypoint tests + fixed a pyhton3 bug
